requirements.txt
tensorflow
matplotlib
ipython
numpy
keras
torch
pandas
scikit-learn
opencv-python
ultralytics

django
djangorestframework
django-cors-headers
requests
mysqlclient
pillow
python-decouple
python-dotenv